requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
psycopg==3.2.9
psycopg-binary==3.2.9
sqlparse==0.5.3
tzdata==2025.2
